Fix empty freestyle-cache-redis distribution (#15)
* Fix empty freestyle-cache-redis distribution.
The freestyle-cache-redis source files were in a `shared` subfolder,
which is only valid for projects cross-built with Scala.js.  Since
this is not the case with freestyle-cache-redis, the artifacts
published from the subproject were empty.  This patch fixes that.

* Fix freestyle-cache-redis source file formatting.
Since freestyle-cache-redis now actually builds, scalafmt also runs
for it.  This patch adds the changes that Scalafmt applies to the
source code when run.

* Fix freestyle-cache-redis test source formatting.

This patch adds the formatting changes that Scalafmt generates when
run on the freestyle-cache-redis test sources.